move language input
