+++ src/backoffice/settings/index.ts
@@ -16,6 +16,7 @@ import { UmbContextProviderController } from '@umbraco-cms/backoffice/context-ap
 import type { UmbEntrypointOnInit } from '@umbraco-cms/backoffice/extensions-api';
 
 import './cultures/components';
+import './languages/components';
 
 export const manifests = [
 	...settingsSectionManifests,
