+++ src/backoffice/settings/languages/components/index.ts
@@ -1,0 +1,1 @@
+import './input-language-picker/input-language-picker.element';
