+++ src/backoffice/settings/languages/workspace/language/views/details/language-details-workspace-view.element.ts
@@ -4,8 +4,9 @@ import { css, html, nothing } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { UmbLanguageWorkspaceContext } from '../../language-workspace.context';
-import { UmbInputCultureSelectElement } from '../../../../../cultures/components/input-culture-select/input-culture-select.element';
-import { UmbInputLanguagePickerElement } from '../../../../../../shared/components/input-language-picker/input-language-picker.element';
+// TODO: set up import alias for these modules
+import type { UmbInputCultureSelectElement } from '../../../../../cultures/components/input-culture-select/input-culture-select.element';
+import type { UmbInputLanguagePickerElement } from '../../../../components/input-language-picker/input-language-picker.element';
 import { UmbChangeEvent } from '@umbraco-cms/backoffice/events';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import { LanguageResponseModel } from '@umbraco-cms/backoffice/backend-api';
