+++ src/backoffice/shared/components/index.ts
@@ -22,7 +22,6 @@ import './extension-slot/extension-slot.element';
 import './input-checkbox-list/input-checkbox-list.element';
 import './input-color-picker/input-color-picker.element';
 import './input-eye-dropper/input-eye-dropper.element';
-import './input-language-picker/input-language-picker.element';
 import './input-media-picker/input-media-picker.element';
 import './input-multi-url-picker/input-multi-url-picker.element';
 import './input-slider/input-slider.element';
